requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
djangorestframework==3.15.2
uvicorn>=0.22.0
dj_database_url==2.2.0
python-dotenv>=0.21
requests
geopy
